Implemented some try-catches to catch the correct exception.

+++ src/main/java/org/yearup/controllers/ShoppingCartController.java
@@ -38,18 +38,20 @@ public class ShoppingCartController
     @PreAuthorize("isAuthenticated()")
     public ShoppingCart getCart(Principal principal)
     {
-        try
-        {
+        try {
             // get the currently logged-in username
             String userName = principal.getName();
             // find database user by userId
             User user = userDao.getByUserName(userName);
+            if(user == null) throw new ResponseStatusException(HttpStatus.NOT_FOUND, "User not found.");
             int userId = user.getId();
             // use the shoppingCartDao to get all items in the cart and return the cart
 //            ShoppingCart test = shoppingCartDao.getByUserId(userId);
 //            System.out.println("Is it true ? " + test.contains(8));
-
             return shoppingCartDao.getByUserId(userId);
+        }
+        catch (ResponseStatusException e) {
+            throw e;
         }
         catch(Exception e)
         {
@@ -66,8 +68,12 @@ public class ShoppingCartController
         try{
             String userName = principal.getName();
             User user = userDao.getByUserName(userName);
+            if(user == null) throw new ResponseStatusException(HttpStatus.NOT_FOUND, "User not found.");
             int userId = user.getId();
             return shoppingCartDao.addProduct(userId, id);
+        }
+        catch (ResponseStatusException e) {
+            throw e;
         } catch (Exception e) {
             throw new ResponseStatusException(HttpStatus.INTERNAL_SERVER_ERROR, "Oops... our bad.");
         }
@@ -83,9 +89,14 @@ public class ShoppingCartController
         try{
             String userName = principal.getName();
             User user = userDao.getByUserName(userName);
+            if(user == null) throw new ResponseStatusException(HttpStatus.NOT_FOUND, "User not found.");
             int userId = user.getId();
             shoppingCartDao.updateCart(userId, id, quantity.getQuantity());
-        } catch (Exception e) {
+        }
+        catch (ResponseStatusException e) {
+            throw e;
+        }
+        catch (Exception e) {
             throw new ResponseStatusException(HttpStatus.INTERNAL_SERVER_ERROR, "Oops... our bad.");
         }
     }
@@ -99,9 +110,14 @@ public class ShoppingCartController
         try{
             String userName = principal.getName();
             User user = userDao.getByUserName(userName);
+            if(user == null) throw new ResponseStatusException(HttpStatus.NOT_FOUND, "User not found.");
             int userId = user.getId();
             shoppingCartDao.deleteCart(userId);
-        } catch (Exception e) {
+        }
+        catch (ResponseStatusException e) {
+            throw e;
+        }
+        catch (Exception e) {
             throw new ResponseStatusException(HttpStatus.INTERNAL_SERVER_ERROR, "Oops... our bad.");
         }
     }
